Removed all references to generic type T for better compatibility with Xamarin.
The original implementation uses reflection to determine the parameterized type at runtime but
this explodes when called from C# because generic types are erased from the generated C#
bindings.

Since this doesn't work in C#, the call to reifyParameterizedTypeClass is replaced with a setter
that explicitly sets the Class of the contained objects.